requirements.txt
Django==1.11.4
djangorestframework==3.6.3
pytz==2017.2
ipdb==0.10.3
djangorestframework-jwt==1.11.0
